sent this version already hope its 90+

+++ src/IDiscountedPriceCalculator.java
@@ -3,5 +3,5 @@ public interface IDiscountedPriceCalculator {
 	
 	public void applyDiscount(int discountPercent);
 	
-	public int getDiscountedPrice();
+	public double getDiscountedPrice();
 }
